# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MassProject API",
        "version": "0.1"
    },
    "paths": {
        "/api/requests": {
            "get": {
                "summary": "Get requests",
                "description": "Get requests with optional parameters",
                "operationId": "d8762545247b9ed012f00e992cbfa665",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "The status of the request",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "active",
                                "resolved"
                            ]
                        }
                    },
                    {
                        "name": "date_from",
                        "in": "query",
                        "description": "The start date",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2022-01-01"
                        }
                    },
                    {
                        "name": "date_to",
                        "in": "query",
                        "description": "The end date",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2023-01-01"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "summary": "Create request",
                "description": "Create request with optional parameters",
                "operationId": "b78212afe2734cba4b079913cc3d3855",
                "parameters": [
                    {
                        "name": "message",
                        "in": "query",
                        "description": "The message of the request",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/requests/{id}": {
            "put": {
                "summary": "Update request comment",
                "description": "Update request comment",
                "operationId": "7ec60803f7bd04fab9b0d7143ad2fe72",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the request",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "comment",
                        "in": "query",
                        "description": "The comment of the request",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Request comment updated successfully"
                    },
                    "404": {
                        "description": "Request not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/user/create": {
            "post": {
                "summary": "Create a new user",
                "description": "Create a new user with required fields",
                "operationId": "fabdea527cb1e495c22eef20c240df05",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "role"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "maxLength": 255
                                    },
                                    "password": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "minLength": 8
                                    },
                                    "role": {
                                        "type": "string",
                                        "enum": [
                                            "user",
                                            "responsible"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter token in format (Bearer <token>)",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}